Chart title, expanding FigureCanvas, and white chart background.
Removed unnecessary autowrap call upon draw event that caused the title
to develop spaces between each character.

Used ax.tick_params() to remove tick marks but still allow a title and
white background. Title is set based on what is being drawn (CPU or RAM
usage). 10% of chart space is used for displaying the title.

With the axes always remaining at a certain aspect ratio, the
FigureCanvas can be expanded to fit the window, so now it is.